fix(mocker): harden LiveEngine lifecycle
